% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{defineMetaSpace}
\alias{defineMetaSpace}
\title{Run single-cell mapping}
\usage{
defineMetaSpace(
  object,
  reference,
  cell_type_var = "annotation_level_4",
  scDF,
  MetaVol = 20
)
}
\arguments{
\item{object}{SPATA2 object}

\item{cell_type_var}{Character value; The col of the Seurat meta data indicating the cell type annotations}

\item{scDF}{Data.frame; Output of the getSingleCellDeconv()}

\item{MetaVol}{Number of genes in meta space}

\item{ref.new}{Seurat Object; Reference dataset used for runRCTD()}
}
\description{
Mapping the output of "getSingleCellDeconv()"to the best matching cell from the reference dataset
}
\author{
Dieter Henrik Heiland
}
