% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{inferCellPositionsfromHE}
\alias{inferCellPositionsfromHE}
\title{Run single-cell mapping}
\usage{
inferCellPositionsfromHE(object, sample.folder)
}
\arguments{
\item{object}{SPATA object}

\item{sample.folder}{path to Visium sample folder}
}
\value{
SPATA object
}
\description{
Mapping the output of "getSingleCellDeconv()"to the best matching cell from the reference dataset
}
\author{
Dieter Henrik Heiland
}
