% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{initiate_Population}
\alias{initiate_Population}
\title{initiate_Population}
\usage{
initiate_Population(
  nr_of_random_spots,
  n_select,
  nested_ref_meta,
  cell_type_var
)
}
\description{
initiate_Population
}
\author{
Dieter Henrik Heiland
}
