% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{initiate_Population_V1}
\alias{initiate_Population_V1}
\title{initiate_Population_V1}
\usage{
initiate_Population_V1(nr_of_random_spots, n_select, nested_ref_meta)
}
\description{
initiate_Population_V1
}
\author{
Dieter Henrik Heiland
}
