% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{runMappingGA}
\alias{runMappingGA}
\title{Run single-cell mapping}
\usage{
runMappingGA(
  object,
  scDF,
  reference,
  cell_type_var,
  metaSpace,
  workers = 16,
  iter = 200,
  iter_GA = 20,
  nr_mut = 2,
  nr_offsprings = 7,
  cross_over_point = 0.5,
  ram = 8
)
}
\arguments{
\item{object}{SPATA2 object}

\item{scDF}{Data.frame; Output of the getSingleCellDeconv()}

\item{reference}{Seurat Object; Reference dataset used for runRCTD()}

\item{cell_type_var}{Character value; The col of the Seurat meta data indicating the cell type annotations}

\item{metaSpace}{The metaSpace definition from the function "defineMetaSpace"}

\item{workers}{Integer, Number of cors}

\item{iter}{Integer: For model BF: Number of random spot compositions; For model oGA size of initial population}

\item{iter_GA}{Integer: Number of iterations of the oGA model}

\item{nr_mut}{Integer: Number of mutations}

\item{nr_offsprings}{Integer: Number of offsprings}

\item{cross_over_point}{Numeric value: Percentage of cross-over cutt-off}

\item{ram}{Integer: GB of ram can be used for multicore session}
}
\description{
Mapping the output of "getSingleCellDeconv()"to the best matching cell from the reference dataset
}
\author{
Dieter Henrik Heiland
}
