% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{runSegmentfromCoords}
\alias{runSegmentfromCoords}
\title{runSegmentfromCoords}
\usage{
runSegmentfromCoords(
  object,
  Coord_file = NULL,
  spot_extension = 0,
  multicore = T,
  workers = 16
)
}
\arguments{
\item{object}{SPATA object}

\item{Coord_file}{data.frame with cell or other coordinates if NULL the SPATAwrappers::getNucleusPosition will look into saves scCoords}

\item{spot_extension}{Extent the area of the spot (from SPATA2 Image processing)}

\item{multicore}{Use multicore !!recommended!!}

\item{workers}{Number of workers for multisession}
}
\value{
data.frame with intensity per spot
}
\description{
Get summary of image intensity per spot
}
\author{
Dieter Henrik Heiland
}
