% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{runSegmentfromImage}
\alias{runSegmentfromImage}
\title{runSegmentfromImage}
\usage{
runSegmentfromImage(
  object,
  Image,
  spot_extension = 0,
  multicore = T,
  workers = 16
)
}
\arguments{
\item{object}{SPATA object}

\item{Image}{EBImage object}

\item{spot_extension}{Extent the area of the spot (from SPATA2 Image processing)}

\item{multicore}{Use multicore !!recommended!!}

\item{workers}{Number of workers for multisession}
}
\value{
data.frame with intensity per spot
}
\description{
Get summary of image intensity per spot
}
\author{
Dieter Henrik Heiland
}
